feat: change to protected path prefix and update middleware
