+++ .env.test
@@ -84,5 +84,5 @@ FEATURE_SYNC_BETAGOUV_USER_API=true
 FEATURE_UNSUBSCRIBE_FROM_INCUBATEUR_MAILING_LIST=true
 FEATURE_USE_NEW_MARRAINAGE=false
 
-# Public API
-PUBLIC_API_KEYS=xxx,yyy
+# Protected API
+PROTECTED_API_KEYS=xxx,yyy
