+++ src/server/config/index.ts
@@ -266,7 +266,11 @@ export default {
     S3_BUCKET: process.env.S3_BUCKET,
     S3_GET_HOST: process.env.S3_GET_HOST,
     S3_REGION: process.env.S3_REGION,
-    // basic protection for public api routes
-    // PUBLIC_API_KEYS: (process.env.PUBLIC_API_KEYS || "").split(",").map(key => key.trim()).filter(Boolean),
-    PUBLIC_API_KEYS: ["secret"], // tmp test
+    protectedAPI: {
+        // basic protection for public api routes
+        // API_KEYS: (process.env.PROTECTED_API_KEYS || "").split(",").map(key => key.trim()).filter(Boolean),
+        API_KEYS: ["secret"], // tmp test
+        paramKeyName: "apiKey" as const,
+        routePrefix: "/api/protected",
+    }
 };
